requirements.txt
mysql-connector-python==9.5.0
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2
prophet==1.2.1
matplotlib==3.10.7
seaborn==0.13.2
scipy==1.16.3
streamlit==1.51.0
plotly==6.4.0
